+++ Sources/CardPayments/CardClient.swift
@@ -130,7 +130,7 @@ public class CardClient: NSObject {
                 }
                 
                 guard let url = url else {
-                    self.notifyFailure(with: CardClientError.missingDeeplinkgURLError)
+                    self.notifyFailure(with: CardClientError.missingDeeplinkURLError)
                     return
                 }